+++ Rasterizer/include/AmberRenderer/Buffers/VertexBuffer.h
@@ -1,4 +1,5 @@
 #pragma once
+
 #include <cstdint>
 
 namespace AmberRenderer::Buffers
