+++ Rasterizer/include/AmberRenderer/Rendering/Settings/EPrimitiveMode.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cstdint>
-
 namespace AmberRenderer::Rendering::Settings
 {
 	enum class EPrimitiveMode : uint8_t
