+++ Rasterizer/include/AmberRenderer/Rendering/Settings/EProjectionMode.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <cstdint>
-
 namespace AmberRenderer::Rendering::Settings
 {
 	enum class EProjectionMode : uint8_t
