+++ Rasterizer/include/AmberRenderer/Rendering/Settings/ERenderingCapability.h
@@ -1,10 +1,8 @@
 #pragma once
 
-#include <cstdint>
-
 namespace AmberRenderer::Rendering::Settings
 {
-	enum ERenderingCapability : uint8_t
+	enum class ERenderingCapability : uint8_t
 	{
 		DEPTH_WRITE = 0x01,
 		DEPTH_TEST = 0x02,
