+++ Rasterizer/include/AmberRenderer/Resources/Material.h
@@ -2,7 +2,7 @@
 
 #include <string>
 
-#include "AmberRenderer/Rendering/Rasterizer/Shaders/AShader.h"
+#include "AmberRenderer/Rendering/SoftwareRenderer/Programs//AProgram.h"
 
 #include "AmberRenderer/Resources/Texture.h"
 
@@ -19,17 +19,17 @@ namespace AmberRenderer::Resources
 
 		bool HasShader() const;
 
-		Rendering::Rasterizer::Shaders::AShader* GetShader() const;
+		Rendering::SoftwareRenderer::Programs::AProgram* GetShader() const;
 		const Texture* GetTexture() const;
 
-		void SetShader(Rendering::Rasterizer::Shaders::AShader* p_shader);
+		void SetShader(Rendering::SoftwareRenderer::Programs::AProgram* p_shader);
 		void SetTexture(Texture* p_texture);
 
 	public:
 		const std::string Name;
 
 	private:
-		Rendering::Rasterizer::Shaders::AShader* m_shader;
+		Rendering::SoftwareRenderer::Programs::AProgram* m_shader;
 		Texture* m_texture;
 	};
 }
