+++ Rasterizer/src/AmberRenderer/Rendering/Driver.cpp
@@ -1,6 +1,6 @@
 #include "AmberRenderer/Rendering/Driver.h"
 
-#include "AmberRenderer/Rendering/Rasterizer/GLRasterizer.h"
+#include "AmberRenderer/Rendering/SoftwareRenderer/AmberGL.h"
 
 AmberRenderer::Rendering::Driver::Driver()
 {
@@ -9,7 +9,7 @@ AmberRenderer::Rendering::Driver::Driver()
 
 void AmberRenderer::Rendering::Driver::SetClearColor(const glm::vec4& p_color) const
 {
-	GLRasterizer::ClearColor(p_color.x, p_color.y, p_color.z, p_color.w);
+	AmberGL::ClearColor(p_color.x, p_color.y, p_color.z, p_color.w);
 }
 
 void AmberRenderer::Rendering::Driver::Clear(bool p_colorBuffer, bool p_depthBuffer) const
@@ -21,55 +21,55 @@ void AmberRenderer::Rendering::Driver::Clear(bool p_colorBuffer, bool p_depthBuf
 
 	if (clearMask != 0)
 	{
-		GLRasterizer::Clear(clearMask);
+		AmberGL::Clear(clearMask);
 	}
 }
 
 void AmberRenderer::Rendering::Driver::SetViewport(uint16_t p_x, uint16_t p_y, uint16_t p_width, uint16_t p_height) const
 {
-	GLRasterizer::Viewport(p_x, p_y, p_width, p_height);
+	AmberGL::Viewport(p_x, p_y, p_width, p_height);
 }
 
 void AmberRenderer::Rendering::Driver::SetRasterizationMode(Settings::ERasterizationMode p_rasterizationMode) const
 {
-	GLRasterizer::PolygonMode(static_cast<uint8_t>(p_rasterizationMode));
+	AmberGL::PolygonMode(static_cast<uint8_t>(p_rasterizationMode));
 }
 
 void AmberRenderer::Rendering::Driver::SetCapability(Settings::ERenderingCapability p_capability, bool p_value) const
 {
-	(p_value ? GLRasterizer::Enable : GLRasterizer::Disable)(static_cast<uint8_t>(p_capability));
+	(p_value ? AmberGL::Enable : AmberGL::Disable)(static_cast<uint8_t>(p_capability));
 }
 
 bool AmberRenderer::Rendering::Driver::GetCapability(Settings::ERenderingCapability p_capability) const
 {
-	return GLRasterizer::IsEnabled(static_cast<uint8_t>(p_capability));
+	return AmberGL::IsEnabled(static_cast<uint8_t>(p_capability));
 }
 
 void AmberRenderer::Rendering::Driver::SetCullFace(Settings::ECullFace p_cullFace) const
 {
-	GLRasterizer::CullFace(static_cast<uint8_t>(p_cullFace));
+	AmberGL::CullFace(static_cast<uint8_t>(p_cullFace));
 }
 
 void AmberRenderer::Rendering::Driver::SetDepthWriting(bool p_enable) const
 {
-	GLRasterizer::DepthMask(p_enable);
+	AmberGL::DepthMask(p_enable);
 }
 
 bool AmberRenderer::Rendering::Driver::GetBool(uint8_t p_parameter) const
 {
 	bool result;
-	GLRasterizer::GetBool(p_parameter, &result);
+	AmberGL::GetBool(p_parameter, &result);
 	return result;
 }
 
 int AmberRenderer::Rendering::Driver::GetInt(uint8_t p_parameter) const
 {
 	int result;
-	GLRasterizer::GetInt(p_parameter, &result);
+	AmberGL::GetInt(p_parameter, &result);
 	return result;
 }
 
 void AmberRenderer::Rendering::Driver::Initialize() const
 {
-	GLRasterizer::Initialize(800, 600);
+	AmberGL::Initialize(800, 600);
 }
