+++ Rasterizer/src/AmberRenderer/Resources/Mesh.cpp
@@ -1,7 +1,5 @@
 #include "AmberRenderer/Resources/Mesh.h"
 
-#include "AmberRenderer/Rendering/Rasterizer/GLRasterizer.h"
-
 AmberRenderer::Resources::Mesh::Mesh(const std::vector<Geometry::Vertex>& p_vertices, const std::vector<uint32_t>& p_indices, uint32_t p_materialIndex) :
 m_vertexCount(static_cast<uint32_t>(p_vertices.size())),
 m_indicesCount(static_cast<uint32_t>(p_indices.size())),
