+++ Rasterizer/src/AmberRenderer/Resources/Texture.cpp
@@ -1,6 +1,6 @@
 #include "AmberRenderer/Resources/Texture.h"
 
-#include "AmberRenderer/Rendering/Rasterizer/GLRasterizer.h"
+#include "AmberRenderer/Rendering/SoftwareRenderer/AmberGL.h"
 
 AmberRenderer::Resources::Texture::Texture(const std::string& p_filePath, uint32_t p_id, uint32_t p_width, uint32_t p_height, uint32_t p_bitsPerPixel, Settings::ETextureFilteringMode p_minFilter, Settings::ETextureFilteringMode p_magFilter, Settings::ETextureWrapMode p_wrapS, Settings::ETextureWrapMode p_wrapT, bool p_generateMipmap) :
 ID(p_id),
@@ -18,11 +18,11 @@ HasMipmaps(p_generateMipmap)
 
 void AmberRenderer::Resources::Texture::Bind(uint32_t p_slot) const
 {
-	GLRasterizer::ActiveTexture(GLR_TEXTURE0 + p_slot);
-	GLRasterizer::BindTexture(GLR_TEXTURE_2D, ID);
+	AmberGL::ActiveTexture(GLR_TEXTURE0 + p_slot);
+	AmberGL::BindTexture(GLR_TEXTURE_2D, ID);
 }
 
 void AmberRenderer::Resources::Texture::Unbind() const
 {
-	GLRasterizer::BindTexture(GLR_TEXTURE_2D, 0);
+	AmberGL::BindTexture(GLR_TEXTURE_2D, 0);
 }
